item
item용 script와 문에 붙이는 코드 정리
