+++ Cave_XP_Collaboration1/Assets/GameManager 1.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 670828891797843430}
   - component: {fileID: 670828891797843431}
   - component: {fileID: 670828891797843428}
+  - component: {fileID: 4377833894520041522}
   m_Layer: 0
   m_Name: GameManager 1
   m_TagString: Untagged
@@ -54,5 +55,18 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 0f2fe2d88f9740b49b70f67259d70d04, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  inventoryPanel: {fileID: 0}
+--- !u!114 &4377833894520041522
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 670828891797843425}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 250f987027857204c83855d5f3b18018, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
