+++ Cave_XP_Collaboration1/Assets/Scenes/Setting_room 1.unity
@@ -215,7 +215,6 @@ RectTransform:
   m_LocalScale: {x: 0, y: 0, z: 0}
   m_Children:
   - {fileID: 578160909}
-  - {fileID: 1482006913}
   m_Father: {fileID: 0}
   m_RootOrder: 24
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -1173,6 +1172,20 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 873a23f0a9a71a44bb78cdf7e840c851, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &1030295744
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1030295741}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 87ee23dd60e5c044f95dd0d26cd6a1af, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  cam0: 0
+  cam1: 0
 --- !u!1001 &1087860043
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1285,6 +1298,10 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: -4620282718447106903, guid: 00026de7bad4c5143a5c830fd3e3deda, type: 3}
+      propertyPath: m_TagString
+      value: ToWare
+      objectReference: {fileID: 0}
     - target: {fileID: 919132149155446097, guid: 00026de7bad4c5143a5c830fd3e3deda, type: 3}
       propertyPath: m_Name
       value: cctv_room
@@ -1305,9 +1322,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 1161795845}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9221adc9ee2764a41957f6ec8b51d212, type: 3}
+  m_Script: {fileID: 11500000, guid: 87ee23dd60e5c044f95dd0d26cd6a1af, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  cam0: 0
+  cam1: 0
 --- !u!65 &1161795847
 BoxCollider:
   m_ObjectHideFlags: 0
@@ -1390,68 +1409,6 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 490d6832ae8241b429e328b75ca068b9, type: 3}
---- !u!1001 &1482006912
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 38916718}
-    m_Modifications:
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_RootOrder
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: -533.11475
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: -300
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8596050509146437130, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_Name
-      value: InvenSet
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
---- !u!4 &1482006913 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-  m_PrefabInstance: {fileID: 1482006912}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1001 &1576831973
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -2043,9 +2000,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 1876902329}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fbd1147060691714490b43257682ca04, type: 3}
+  m_Script: {fileID: 11500000, guid: 87ee23dd60e5c044f95dd0d26cd6a1af, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  cam0: 0
+  cam1: 0
 --- !u!65 &1876902331
 BoxCollider:
   m_ObjectHideFlags: 0
