+++ Cave_XP_Collaboration1/Assets/Scenes/WarehouseScene.unity
@@ -919,7 +919,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 681896185}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a4089bd55de189f4aa56fda8df8dc0c3, type: 3}
+  m_Script: {fileID: 11500000, guid: 87ee23dd60e5c044f95dd0d26cd6a1af, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!65 &681896188
@@ -1100,7 +1100,6 @@ RectTransform:
   m_LocalScale: {x: 0, y: 0, z: 0}
   m_Children:
   - {fileID: 1533684394}
-  - {fileID: 987728708}
   m_Father: {fileID: 0}
   m_RootOrder: 36
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -1494,68 +1493,6 @@ PrefabInstance:
       objectReference: {fileID: 2100000, guid: d219a393cd0d8e5468ff5cf7f639461d, type: 2}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: e43edfde475fd8247bf05c31b3753431, type: 3}
---- !u!1001 &987728707
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 804897654}
-    m_Modifications:
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_RootOrder
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: -533.11475
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: -300
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8596050509146437130, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-      propertyPath: m_Name
-      value: InvenSet
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
---- !u!4 &987728708 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 7064144408758427896, guid: 573f9af711306d44787eab7cbaa16682, type: 3}
-  m_PrefabInstance: {fileID: 987728707}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1001 &1012387591
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -2672,6 +2609,10 @@ PrefabInstance:
       propertyPath: m_Materials.Array.data[0]
       value: 
       objectReference: {fileID: 2100000, guid: 1962cd9b592d6ea46b2e85bb8bc4d471, type: 2}
+    - target: {fileID: -4620282718447106903, guid: edef2fefc1e724f4abcbb07b660cfc33, type: 3}
+      propertyPath: m_TagString
+      value: ToSetting
+      objectReference: {fileID: 0}
     - target: {fileID: -4509823143501964880, guid: edef2fefc1e724f4abcbb07b660cfc33, type: 3}
       propertyPath: m_Materials.Array.data[0]
       value:
