+++ Cave_XP_Collaboration1/Assets/Scripts/DoorAndGame.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 873a23f0a9a71a44bb78cdf7e840c851
+guid: 87ee23dd60e5c044f95dd0d26cd6a1af
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
