--- Cave_XP_Collaboration1/Assets/Scripts/Quit.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 250f987027857204c83855d5f3b18018
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
