--- Cave_XP_Collaboration1/Assets/Scripts/Touch.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b17e88250fdc9ee43aa3a892ec6a82ac
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
