--- Cave_XP_Collaboration1/Assets/Scripts/Touch/ControlToSetting.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9d5347e6e2f38c54f9a36a42bb442a14
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
