--- Cave_XP_Collaboration1/Assets/Scripts/Touch/MonsterRoom.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 620daacdb48c8134688d2f56a0e1a75b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
