--- Cave_XP_Collaboration1/Assets/Scripts/Touch/SettingToWare.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9221adc9ee2764a41957f6ec8b51d212
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
