--- Cave_XP_Collaboration1/Assets/Scripts/Touch/TouchPuzzle2.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class TouchPuzzle2 : MonoBehaviour
-{
-    private void OnMouseDown()
-    {
-        Debug.Log("¼¼ÆÃ·ë °¡");
-        SceneManager.LoadScene("MemoryScene");
-    }
-}
-
-
