--- Cave_XP_Collaboration1/Assets/Scripts/Touch/TouchPuzzle3.cs
@@ -1,14 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class TouchPuzzle3 : MonoBehaviour
-{
-   
-    private void OnMouseDown()
-    {
-        Debug.Log("¼¼ÆÃ·ë °¡");
-        SceneManager.LoadScene("AnagramGame");
-    }
-}
