--- Cave_XP_Collaboration1/Assets/Scripts/Touch/TouchPuzzle3.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fbd1147060691714490b43257682ca04
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
