--- Cave_XP_Collaboration1/Assets/Scripts/Touch/WareToSetting.cs
@@ -1,13 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class WareToSetting : MonoBehaviour
-{
-    private void OnMouseDown()
-    {
-        Debug.Log("¼¼ÆÃ·ë °¡");
-        SceneManager.LoadScene("Setting_room 1");
-    }
-}
