--- Cave_XP_Collaboration1/Assets/Scripts/Touch/WareToSetting.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a4089bd55de189f4aa56fda8df8dc0c3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
