+++ Cave_XP_Collaboration1/UserSettings/EditorUserSettings.asset
@@ -24,16 +24,16 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c72151b180a331f2b233e2867083debf42d
       flags: 0
     RecentlyUsedScenePath-6:
-      value: 22424703114646680e0b0227036c78111b122c2f3f3c15332827187df7ee3d2cfb
-      flags: 0
-    RecentlyUsedScenePath-7:
-      value: 22424703114646680e0b0227036c7c1f18030a252024193d22270e27edf20b2aede83ba7f234362820
-      flags: 0
-    RecentlyUsedScenePath-8:
       value: 22424703114646680e0b0227036c6c15020311242b17343f22245d62acf53a31f6fe
       flags: 0
-    RecentlyUsedScenePath-9:
+    RecentlyUsedScenePath-7:
       value: 22424703114646680e0b0227036c681104121025393b23032e2c1336acf53a31f6fe
+      flags: 0
+    RecentlyUsedScenePath-8:
+      value: 22424703114646680e0b0227036c78111b122c2f3f3c15332827187df7ee3d2cfb
+      flags: 0
+    RecentlyUsedScenePath-9:
+      value: 22424703114646680e0b0227036c7c1f18030a252024193d22270e27edf20b2aede83ba7f234362820
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
